Finished tests in MySqlAuthDaoTests

+++ server/src/main/java/dataaccess/sql/MySqlAuthDao.java
@@ -72,6 +72,9 @@ public class MySqlAuthDao implements AuthDAO {
     }
 
     public void deleteAuth(String authToken) throws DataAccessException {
+        if (authToken == null) {
+            throw new DataAccessException("Invalid request");
+        }
         try (var conn = DatabaseManager.getConnection()) {
             var statement = "DELETE FROM auth WHERE authToken = ?";
             try (var preparedStatement = conn.prepareStatement(statement)) {
@@ -84,6 +87,9 @@ public class MySqlAuthDao implements AuthDAO {
     }
 
     public String getUser(String authToken) throws DataAccessException {
+        if (authToken == null) {
+            throw new DataAccessException("Invalid request");
+        }
         try (var conn = DatabaseManager.getConnection()) {
             var statement = "SELECT username FROM auth WHERE authToken = ?";
             try (var preparedStatement = conn.prepareStatement(statement)) {
